ocr_agent/requirements.txt
# ocr_agent/requirements.txt
pytesseract
pillow
